+++ eulfedora/views.py
@@ -35,6 +35,7 @@ import logging
 from django.contrib.auth import views as authviews
 from django.http import HttpResponse, Http404, HttpResponseBadRequest
 from django.views.decorators.http import require_http_methods, condition
+from django.views.generic import View
 
 from eulfedora.util import RequestFailed
 from eulfedora.server import Repository, FEDORA_PASSWORD_SESSION_KEY
@@ -96,6 +97,7 @@ def datastream_lastmodified(request, pid, dsid, type=None, repo=None,
             return ds.created
     except RequestFailed:
         pass
+
 
 
 @condition(etag_func=datastream_etag)
@@ -413,3 +415,75 @@ def login_and_store_credentials_in_session(request, *args, **kwargs):
         # on successful login, encrypt and store user's password to use for fedora access
         request.session[FEDORA_PASSWORD_SESSION_KEY] = encrypt(request.POST.get('password'))
     return response
+
+
+
+# class-based views
+
+class RawDatastreamView(View):
+    '''Class-based view for serving out datastream content from Fedora.
+    '''
+    #: subclass of DigitalObject, if needed
+    object_type = None
+    #: datastream id
+    datastream_id = ''
+    #: Enable range requests (default: False)
+    accept_range_request = False
+    #: url kwarg term for retrieving object pid (default: pid)
+    pid_url_kwarg = 'pid'
+    #: Repository class to use, if needed
+    repository_class = Repository
+    #: extra http headers to include
+    headers = {}
+
+    @classmethod
+    def etag(cls, request, *args, **kwargs):
+        '''Class method to generate an ETag for use with
+        conditional processing; calls :meth:`datastream_etag` with
+        class configuration.'''
+        pid = kwargs[cls.pid_url_kwarg]
+        return datastream_etag(request, pid, cls.datastream_id,
+            type=cls.object_type, repo=cls.repository_class(request=request),
+                    accept_range_request=cls.accept_range_request)
+
+    @classmethod
+    def last_modified(cls, request, *args, **kwargs):
+        '''Class method to generate last-modified header for use with
+        conditional processing; calls :meth:`datastream_lastmodified` with
+        class configuration.'''
+        pid = kwargs[cls.pid_url_kwarg]
+        return datastream_lastmodified(request, pid, cls.datastream_id,
+            type=cls.object_type, repo=cls.repository_class(request=request),
+                    accept_range_request=cls.accept_range_request)
+
+    @classmethod
+    def as_view(cls, **initkwargs):
+        view = super(RawDatastreamView, cls).as_view(**initkwargs)
+        # wrap view with conditional decorator for etag/last-modified
+        return condition(etag_func=cls.etag,
+            last_modified_func=cls.last_modified)(view)
+
+    def get_datastream_id(self):
+        return self.datastream_id
+
+    def get_repository(self):
+        '''Initialize and return the configured repository class,
+        passing in the current request.'''
+        return self.repository_class(request=self.request)
+
+    def get_headers(self):
+        '''Return headers to be included when generating the datastream
+        content response.  Default implementation is to return
+        :attr:`headers`.'''
+        return self.headers
+
+    def head(self, request, *args, **kwargs):
+        # raw_datastream method handles both head and get
+        return self.get(request, *args, **kwargs)
+
+    def get(self, request, *args, **kwargs):
+        pid = kwargs[self.pid_url_kwarg]
+        return raw_datastream(request, pid, self.get_datastream_id(),
+            type=self.object_type, repo=self.get_repository(),
+            headers=self.get_headers(),
+            accept_range_request=self.accept_range_request)